// packages/openapi/openapi.ts
export interface MediaTypeObject {
  schema?: SchemaObject | ReferenceObject;
  example?: unknown;
  examples?: { [name: string]: ExampleObject | ReferenceObject };
  encoding?: { [encoding: string]: EncodingObject };
}

export interface ExampleObject {
  summary?: string;
  description?: string;
  value?: unknown;
  externalValue?: string;
}
// ...
export type SchemaObject = Schema;

export interface DiscriminatorObject {
  propertyName: string;
  mapping?: { [value: string]: string };
}

export interface XMLObject {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;
}

export interface ExternalDocsObject {
  url: string;
  description?: string;
}

// Reference Object
export interface ReferenceObject {
  $ref: string;
  summary?: string; // OpenAPI 3.1
  description?: string; // OpenAPI 3.1
}

// Tag Object
export interface TagObject {
  name: string;
  description?: string;
  externalDocs?: ExternalDocsObject;
}

// Header Object
export interface HeaderObject {
  description?: string;
  required?: boolean;
  deprecated?: boolean;
  allowEmptyValue?: boolean;
  style?: string;
  explode?: boolean;
  allowReserved?: boolean;
  schema?: SchemaObject | ReferenceObject;
  example?: unknown;
  examples?: { [example: string]: ExampleObject | ReferenceObject };
  content?: { [mediaType: string]: MediaTypeObject };
}

// Encoding Object
export interface EncodingObject {
  contentType?: string;
  headers?: { [header: string]: HeaderObject | ReferenceObject };
  style?: string;
  explode?: boolean;
  allowReserved?: boolean;
}